split base in response, and a real base which is also used by error
